Adjust `knit_print()` output of `StoppingList`
This patch adjust the preamble of `knit_print()` output for `StoppingList`,
`StoppingAny` and `StoppingAll` according to the length of stop list.

For `StoppingList`, the premable is adjusted according to whether the length of
stop list is 1 or higher. And the wording "rule" or "rules" are used.

For `StoppingAny` and `StoppingAll`, the preamble is adjusted accroding to the
length of stop list being 1, 2 or more. And the wording "this"/"either"/"any" or
"this"/"both"/"all" and "rule"/"rules" are used accordingly.